[CLEANUP] Refactored and organized code
